+++ src/main/java/magazin/demo/dto/MonitorsDTO.java
@@ -9,7 +9,7 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 public class MonitorsDTO {
 
-    private int serialNumber;
+    private long serialNumber;
     private String manufacturer;
     private int price;
     private int quantity;
